requirements.txt
PyStemmer==2.0.1